requirements.txt
# asgiref==3.2.10
# boto3==1.14.61
# botocore==1.17.61
# certifi==2020.6.20
# cffi==1.14.3
# chardet==3.0.4
# cryptography==3.1.1
# defusedxml==0.7.0rc1
# dj-database-url==0.5.0
# dj-static==0.0.6
# Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-markdownx==3.0.1
django-storages==1.10.1
django-summernote==0.8.11.6
# docutils==0.15.2
# gunicorn==20.0.4
# idna==2.10
# importlib-metadata==3.1.1
# jmespath==0.10.0
Markdown==3.3.3
# oauthlib==3.1.0
# olefile==0.46
# Pillow @ file:///C:/ci/pillow_1594298234712/work
# pycparser==2.20
# PyJWT==1.7.1
# python-dateutil==2.8.1
# python-social-auth==0.3.6
# python3-openid==3.2.0
# pytz==2020.1
# requests==2.24.0
# requests-oauthlib==1.3.0
# s3transfer==0.3.3
# six==1.15.0
# social-auth-app-django==4.0.0
# social-auth-core==3.3.3
# sqlparse==0.3.1
# static3==0.7.0
# stripe==2.55.0
# urllib3==1.25.10
# whitenoise==5.2.0
# wincertstore==0.2
# zipp==3.4.0

asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
gunicorn==20.0.4
idna==2.8
Pillow==5.4.1
psycopg2-binary==2.8.2
pytz==2020.1
requests==2.21.0
setuptools==49.2.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.24.1
whitenoise==5.2.0
django-storages==1.10.1
boto3==1.14.61 
botocore==1.17.61 
docutils==0.15.2 
jmespath==0.10.0 
python-dateutil==2.8.1 
s3transfer==0.3.3 
six==1.15.0 
